+++ plone/app/contenttypes/migration/utils.py
@@ -119,8 +119,11 @@ def _compareSchemata(interface):
     portal = getSite()
     pc = portal.portal_catalog
     brains = pc(object_provides=interface.__identifier__)
-    if brains:
-        obj = brains[0].getObject()
+    for brain in brains:
+        if not brain.meta_type or 'dexterity' in brain.meta_type.lower():
+            # There might be DX types with same iface and meta_type than AT
+            continue
+        obj = brain.getObject()
         real_fields = set(obj.Schema()._names)
         orig_fields = set(obj.schema._names)
         diff = [i for i in real_fields.difference(orig_fields)]