+++ CHANGES.rst
@@ -4,11 +4,35 @@ Changelog
 1.2a2 (unreleased)
 ------------------
 
+- Do a better check, if LinguaPlone is installed, based on the presence of the
+  "LinguaPlone" browser layer. Asking the quick installer tool might claim it's
+  installed, where it's not.
+  [thet]
+
+- Register folderish views not for plone.app.contenttypes' IFolder but for
+  plone.dexterity's IDexterityContainer. Now, these views can be used on any
+  folderish Dexterity content.
+  [thet]
+
+- Add a ICustomMigrator interface to the migration framework, which can be used
+  to register custom migrator adapters. This can be useful to add custom
+  migrators to more than one or all content types. For example for
+  schemaextenders, which are registered on a interface, which is provided by
+  several content types.
+  [thet]
+
+- In the migration framework, fix queries for Archetype objects, where only
+  interfaces are used to skip brains with no or Dexterity meta_type. In some
+  cases Dexterity and Archetype objects might provide the same marker
+  interfaces.
+  [thet]
+
 - Add logging messages to content migrator for more verbosity on what's
   happening while running the migration.
   [thet]
 
-- Fix @@atct_migrator and @@atct_migrator_results templates work within Plone3.
+- Use Plone 4 based @@atct_migrator and @@atct_migrator_results template
+  structure.
   [thet]